Add (G49): Update rate artwork feat

+++ src/Core/Domain/PenomyAPI.Domain.RelationalDb/Repositories/Features/Generic/IG49Repository.cs
@@ -5,6 +5,6 @@ namespace PenomyAPI.Domain.RelationalDb.Repositories.Features.Generic;
 
 public interface IG49Repository
 {
-    Task<bool> IsArtworkExistsAsync(int id, CancellationToken cancellationToken);
+    Task<bool> IsArtworkExistsAsync(long id, CancellationToken cancellationToken);
     Task<bool> RateArtworkAsync(long userId, long artworkId, byte starRates, CancellationToken cancellationToken);
 }
